pyproject.toml
[project]
name = "opengraphau"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gdown>=5.2.0",
    "pillow>=11.3.0",
    "timm>=1.0.19",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
]
license = { file = "LICENSE" }

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build]

[tool.hatch.build.targets.wheel]
packages = ["opengraphau"]
